Separated the handling of an event and the logic.
What
===
Separated the handling of the click event on the Verses element and the
business logic for toggling verses.

Why
===
Separation of concerns and single responsibility best practices. There
was a function that was an event handler and dealt with suppressing the
default action of clicking an anchor, which also handled the logic for
toggling verses.