Have the current day div take up space initially.
What
===
Changed the default contents of the date in the navigation so that it
takes up the same whitespace as when it will have text in it.

Why
===
Otherwise the height of the date div changes when the date is inserted,
causing the rest of the page contents to move.