Tidied the index.html inline JavaScript.
What
===
Tidied the index.html inline JavaScript by placing code that was
duplicated multiple times in the page into functions and calling them.

Why
===
Reduce the size of the resulting HTML and generally nicer looking code.

Considerations
===
I considered moving this code into an `index.js`, but I want this
JavaScript to run as the page renders, not after the page renders. At
some point I may look into moving it into a file and directing the
template to inline the file into the page instead.