Time to complete including video and reading.
What
===
Moved the reading time to the top and expanded it to include the video
and all reading in a single estimate.

Why
===
It's more useful to know upfront how much time we need to allocate at
least to the reading. This helps make sure we give ourselves adequate
time where we aren't interrupted.